Implement ListViewModel

+++ SwiftPhotoBoardExample/ListView/ListViewModel.swift
@@ -10,25 +10,35 @@ import SwiftUI
 
 @Observable
 final class ListViewModel {
-    private let modelContext: ModelContext
+    private let modelContext: ModelContext?
     private let thumbnailService: ThumbnailService
 
     init(
         modelContext: ModelContext,
-        thumbnailStore: ThumbnailService = ThumbnailServiceImpl()
+        thumbnailService: ThumbnailService? = nil,
     ) {
         self.modelContext = modelContext
-        self.thumbnailService = thumbnailStore
+        self.thumbnailService = thumbnailService ?? ThumbnailServiceImpl()
     }
 
-    func deleteItems(items: [Item], offsets: IndexSet) {
+    // Dummy for initialization; not actually used.
+    init() {
+        self.modelContext = nil
+        self.thumbnailService = ThumbnailServiceImpl()
+    }
+
+    func deleteItems(items: [Item], offsets: IndexSet) throws {
+        guard let modelContext else { throw SwiftDataError.missingModelContext }
+
         for index in offsets {
             let item = items[index]
             if let localIdentifier = item.localIdentifier {
-                thumbnailService.deleteThumbnail(localIdentifier: localIdentifier)
+                thumbnailService.deleteThumbnail(
+                    localIdentifier: localIdentifier
+                )
             }
             modelContext.delete(item)
-            try? modelContext.save()
+            try modelContext.save()
         }
     }
 }
